test(dev): CTL-1216 — execution-core fixtures onto the leaf, and doctor's window moved out of doctor.mjs
The last tail, and the one my own local run had already told me about — 195
failures against an 80 baseline — before I pushed. I read the file's tail instead
of its summary. CI then said the same thing, more expensively.

25 execution-core test files derived the event-log filename themselves and so
addressed a monthly file the now-weekly code never opens. All repointed at the
production leaf.

TWO SELF-INFLICTED DEFECTS, both worth recording because both were found by
something failing LOUDLY rather than by review:

1. Wrong import depth, for the third time in this branch. `execution-core/*.test.mjs`
   needs `../lib/`, not `../../lib/`. My helper guessed depth from a path shape
   again and got it inverted, and 7 files failed to load — which silently
   shortened the run to 9,731 tests from 11,603 while REDUCING the failure count
   to 75, i.e. it looked like an improvement. Import paths are now COMPUTED with
   os.path.relpath against the real target and every one is asserted to resolve
   before the suite runs. "Fewer failures" is not the metric; the test count is.

2. `doctor.mjs` gained 19 CTL-1529 whole-file-read-guard violations. None of them
   read the event log — they read launchd plists, .zshenv, env files. That guard
   propagates event-log taint by NAME and dataflow, so introducing an
   `eventsDir`-shaped local into a 6,000-line file taints every downstream
   `readFileSync(p, …)` in it. `origin/main`'s doctor.mjs has ZERO hits; mine had
   19.

   The tempting fix was to rename the local until the name-based heuristic
   stopped matching. That is exactly the evasion that guard's own header forbids
   ("no one can self-exempt"). Instead the window resolver moves to
   execution-core/doctor-event-log-window.mjs — a file that genuinely deals with
   event-log paths and reads nothing at all. doctor.mjs returns to zero, which is
   where it was before this ticket, and the taint sits where it belongs.

On flakiness, measured rather than asserted: two IDENTICAL full-suite runs of
execution-core give 80 and 87 failures with 7 differing names, and the
`reclaimDeadWorkIfPossible` family that appears there passes 347/0 in isolation
and is absent entirely from a run of CI's actual allowlist. Those are full-suite
ordering artifacts, not this diff.

Against CI's exact 272-file allowlist: 9,888 pass / 50 fail, and the 50 are the
pre-existing `runCli` spawnSync-surface / boot-drain / pino-fallback set that
also failed on the recorded pre-flip baseline.

Refs CTL-1216.
