+++ plugins/dev/scripts/execution-core/daemon-watchdog-predicates.test.mjs
@@ -25,6 +25,10 @@ import {
   forwarderEventLogPath,
   DAEMON_WATCHDOG_TARGETS,
 } from "./daemon-watchdog-predicates.mjs";
+// CTL-1216: resolve the event-log filename through the production leaf so this
+// fixture follows the ACTIVE scheme. A pinned monthly name addresses a file the
+// code under test never opens.
+import { eventLogBasenameFor, resolveRotationScheme } from "../lib/event-log-paths.mjs";
 
 function tmp() {
   const dir = mkdtempSync(join(tmpdir(), "dw-pred-"));
@@ -381,7 +385,7 @@ describe("forwarderEventLogPath (matches otel-forward's tail target)", () => {
 
   function ym() {
     const now = new Date();
-    return `${now.getUTCFullYear()}-${String(now.getUTCMonth() + 1).padStart(2, "0")}`;
+    return eventLogBasenameFor(now, resolveRotationScheme({ env: process.env })).replace(/\.jsonl$/, "");
   }
 
   test("defaults to $CATALYST_DIR/events/<UTC-month>.jsonl", () => {
