+++ plugins/dev/scripts/execution-core/delegate-config-event.test.mjs
@@ -22,6 +22,10 @@ import {
   readDelegateQueueDepth,
   readGovernanceConfig,
 } from "./config.mjs";
+// CTL-1216: resolve the event-log filename through the production leaf so this
+// fixture follows the ACTIVE scheme. A pinned monthly name addresses a file the
+// code under test never opens.
+import { eventLogBasenameFor, resolveRotationScheme } from "../lib/event-log-paths.mjs";
 
 // ---------------------------------------------------------------------------
 // appendDispatchEnqueuedEvent — round-trip (mirrors the CTL-660 dispatch block)
@@ -44,7 +48,7 @@ describe("appendDispatchEnqueuedEvent (CTL-1331 creation telemetry)", () => {
   // Read back the single envelope written this test (current UTC month log).
   function readBackEnvelope() {
     const now = new Date();
-    const ym = `${now.getUTCFullYear()}-${String(now.getUTCMonth() + 1).padStart(2, "0")}`;
+    const ym = eventLogBasenameFor(now, resolveRotationScheme({ env: process.env })).replace(/\.jsonl$/, "");
     const lines = readFileSync(join(envCatalystDir, "events", `${ym}.jsonl`), "utf8")
       .split("\n")
       .filter(Boolean);
