+++ plugins/dev/scripts/execution-core/integration-ctl-705.test.mjs
@@ -21,6 +21,10 @@ import {
 } from "./scheduler.mjs";
 import { log } from "./config.mjs";
 import { ownedBy } from "./hrw.mjs";
+// CTL-1216: resolve the event-log filename through the production leaf so this
+// fixture follows the ACTIVE scheme. A pinned monthly name addresses a file the
+// code under test never opens.
+import { eventLogBasenameFor, resolveRotationScheme } from "../lib/event-log-paths.mjs";
 
 let orchDir;
 let catalystDir;
@@ -634,7 +638,7 @@ describe("CTL-705 acceptance scenario — preemption + resume", () => {
 
     // Verify the event log contains a preemption event for CTL-2
     const now2 = new Date();
-    const ym = `${now2.getUTCFullYear()}-${String(now2.getUTCMonth() + 1).padStart(2, "0")}`;
+    const ym = eventLogBasenameFor(now2, resolveRotationScheme({ env: process.env })).replace(/\.jsonl$/, "");
     const eventLog = readFileSync(join(catalystDir, "events", `${ym}.jsonl`), "utf8")
       .split("\n")
       .filter(Boolean)
