+++ plugins/dev/scripts/execution-core/integration.test.mjs
@@ -33,6 +33,10 @@ import {
 } from "./scheduler.mjs";
 import { recoverStartup, defaultStatJob } from "./recovery.mjs";
 import { loadCursor } from "./event-cursor.mjs";
+// CTL-1216: resolve the event-log filename through the production leaf so this
+// fixture follows the ACTIVE scheme. A pinned monthly name addresses a file the
+// code under test never opens.
+import { eventLogBasenameFor, resolveRotationScheme } from "../lib/event-log-paths.mjs";
 
 let catalystDir;
 let prevCatalystDir;
@@ -204,7 +208,7 @@ describe("execution-core integration — crash recovery (CTL-539)", () => {
   // appendEventLog — append a line to the current UTC month's event log.
   function eventLogPath() {
     const now = new Date();
-    const ym = `${now.getUTCFullYear()}-${String(now.getUTCMonth() + 1).padStart(2, "0")}`;
+    const ym = eventLogBasenameFor(now, resolveRotationScheme({ env: process.env })).replace(/\.jsonl$/, "");
     return join(catalystDir, "events", `${ym}.jsonl`);
   }
   function appendEventLog(line) {
