+++ plugins/dev/scripts/execution-core/lib/dual-envelope.test.mjs
@@ -24,6 +24,10 @@ import { join } from "node:path";
 import { buildDualEnvelopeLine, FLAT_ATTRIBUTE_MAP } from "./canonical-event.mjs";
 import { getEventName } from "../../lib/event-name.mjs"; // CTL-1834: moved out of broker/
 import { isFlatEvent, isPinoRecord } from "../../otel-forward/lib/normalize.ts";
+// CTL-1216: resolve the event-log filename through the production leaf so this
+// fixture follows the ACTIVE scheme. A pinned monthly name addresses a file the
+// code under test never opens.
+import { eventLogBasenameFor, resolveRotationScheme } from "../../lib/event-log-paths.mjs";
 
 const parse = (line) => JSON.parse(line.trimEnd());
 
@@ -168,7 +172,7 @@ describe("every live JS v1 emit site now writes the superset line", () => {
 
   const readSoleEvent = () => {
     const now = new Date();
-    const ym = `${now.getUTCFullYear()}-${String(now.getUTCMonth() + 1).padStart(2, "0")}`;
+    const ym = eventLogBasenameFor(now, resolveRotationScheme({ env: process.env })).replace(/\.jsonl$/, "");
     const path = join(dir, "events", `${ym}.jsonl`);
     expect(existsSync(path)).toBe(true);
     const lines = readFileSync(path, "utf8").trim().split("\n");
