+++ plugins/dev/scripts/execution-core/monitor.coordination.test.mjs
@@ -33,6 +33,10 @@ import {
 import { log } from "./config.mjs";
 import { dropProject } from "./eligible-set.mjs";
 import { __resetFleetFreezeLatch } from "./fleet-freeze-alert.mjs";
+// CTL-1216: resolve the event-log filename through the production leaf so this
+// fixture follows the ACTIVE scheme. A pinned monthly name addresses a file the
+// code under test never opens.
+import { eventLogBasenameFor, resolveRotationScheme } from "../lib/event-log-paths.mjs";
 
 // ── test harness ─────────────────────────────────────────────────────────────
 
@@ -82,7 +86,7 @@ function appendCoordination(obj) {
 // tail alongside the coordination tail for the shared cross-source dedup.
 function eventLogPath() {
   const now = new Date();
-  const ym = `${now.getUTCFullYear()}-${String(now.getUTCMonth() + 1).padStart(2, "0")}`;
+  const ym = eventLogBasenameFor(now, resolveRotationScheme({ env: process.env })).replace(/\.jsonl$/, "");
   return join(catalystDir, "events", `${ym}.jsonl`);
 }
